Add files via upload
The objective is figuring out the dynamics that influence where and when fires will occur to help authorities better understand their effects. The challenge was to predict how these dynamics will play out in the future, under different climatic conditions and provide the predicted amount of burnt area in a region.